# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Face recognition search API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LibraryImage {
  id: number;
  filename: string;
  originalName: string;
  faceCount: number;
  uploadedAt: string;
  url: string;
}

export interface ImageUploadInput {
  /** Image file (sent as multipart/form-data) */
  file: string;
}

export interface SearchInput {
  /** Query image file (sent as multipart/form-data) */
  file: string;
  /** Cosine similarity threshold (0.0-1.0, default 0.40) */
  threshold?: number;
}

export interface SearchMatch {
  imageId: number;
  filename: string;
  originalName: string;
  url: string;
  /** Cosine similarity score (0-1) */
  score: number;
  faceCount: number;
  uploadedAt: string;
}

export interface SearchResult {
  matches: SearchMatch[];
  queryFacesFound: number;
  totalSearched: number;
}

export interface LibraryStats {
  totalImages: number;
  totalFaces: number;
  embeddingsStored: number;
}

export interface ErrorResponse {
  error: string;
}