# lib/api-zod/src/generated/types/imageUploadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Face recognition search API
 * OpenAPI spec version: 0.1.0
 */

export interface ImageUploadInput {
  /** Image file (sent as multipart/form-data) */
  file: string;
}