# lib/api-zod/src/generated/types/searchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Face recognition search API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchInput {
  /** Query image file (sent as multipart/form-data) */
  file: string;
  /** Cosine similarity threshold (0.0-1.0, default 0.40) */
  threshold?: number;
}