# lib/api-zod/src/generated/types/searchMatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Face recognition search API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchMatch {
  imageId: number;
  filename: string;
  originalName: string;
  url: string;
  /** Cosine similarity score (0-1) */
  score: number;
  faceCount: number;
  uploadedAt: string;
}