# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Face recognition search API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchMatch } from './searchMatch';

export interface SearchResult {
  matches: SearchMatch[];
  queryFacesFound: number;
  totalSearched: number;
}